pyproject.toml
[project]
name = "autogenmeeting1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyautogen>=0.2.0",
    "python-dotenv>=1.0.0",
    "autogen-ext[openai]",
]
